+++ generators/app/inputs/edit.js
@@ -20,6 +20,7 @@
 	const color          = require('./color.js');
 	const lang           = require('./lang.js');
 	const theme          = require('./theme.js');
+	const twitterName    = require('./twitterName.js');
 
 	const longerGenerator = 20;
 	const prompts         = [
@@ -68,6 +69,10 @@
 				{
 					value: 'aalBowerDependency',
 					short: 'aalBowerDependency'
+				},
+				{
+					value: 'twitterName',
+					short: 'twitterName'
 				}
 			]
 		}
@@ -84,16 +89,17 @@
 			logs.hintsMultiline($that, hints);
 
 			/* eslint no-magic-numbers:"off" */
-			prompts[0].choices[0].name = logs.choice('App name', $that.appName, longerGenerator);
-			prompts[0].choices[1].name = logs.choice('App description', $that.appDescription, longerGenerator);
-			prompts[0].choices[2].name = logs.choice('Theme', $that.theme, longerGenerator);
-			prompts[0].choices[3].name = logs.choice('Lang', $that.lang, longerGenerator);
-			prompts[0].choices[4].name = logs.choice('Background color', $that.backgroundColor, longerGenerator);
-			prompts[0].choices[5].name = logs.choice('Theme color', $that.themeColor, longerGenerator);
-			prompts[0].choices[6].name = logs.choice('Author first name', $that.authorFirstName, longerGenerator);
-			prompts[0].choices[7].name = logs.choice('Author last name', $that.authorLastName, longerGenerator);
-			prompts[0].choices[8].name = logs.choice('Author email', $that.authorEmail, longerGenerator);
-			prompts[0].choices[9].name = logs.choice('AAL bower dependency', $that.aal, longerGenerator);
+			prompts[0].choices[0].name  = logs.choice('App name', $that.appName, longerGenerator);
+			prompts[0].choices[1].name  = logs.choice('App description', $that.appDescription, longerGenerator);
+			prompts[0].choices[2].name  = logs.choice('Theme', $that.theme, longerGenerator);
+			prompts[0].choices[3].name  = logs.choice('Lang', $that.lang, longerGenerator);
+			prompts[0].choices[4].name  = logs.choice('Background color', $that.backgroundColor, longerGenerator);
+			prompts[0].choices[5].name  = logs.choice('Theme color', $that.themeColor, longerGenerator);
+			prompts[0].choices[6].name  = logs.choice('Author first name', $that.authorFirstName, longerGenerator);
+			prompts[0].choices[7].name  = logs.choice('Author last name', $that.authorLastName, longerGenerator);
+			prompts[0].choices[8].name  = logs.choice('Author email', $that.authorEmail, longerGenerator);
+			prompts[0].choices[9].name  = logs.choice('AAL bower dependency', $that.aal, longerGenerator);
+			prompts[0].choices[10].name = logs.choice('Twitter name', $that.twitterName, longerGenerator);
 
 			$that.prompt(prompts).then($response => {
 				Edit.onSuccess($that, $response, deferred);
@@ -133,6 +139,9 @@
 					break;
 				case 'aal':
 					$deferred.resolve(aal.set($that, true));
+					break;
+				case 'twitterName':
+					$deferred.resolve(twitterName.set($that, true));
 					break;
 				default:
 					break;